rec-service/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
psycopg2-binary
python-dotenv
pydantic